// frontend/src/api/model/userCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface UserCreate {
  /**
     * @minLength 3
     * @maxLength 180
     */
  email: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  full_name: string;
  role: UserRole;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
  is_active?: boolean;
}