// frontend/src/api/model/userRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface UserRead {
  id: string;
  email: string;
  full_name: string;
  role: UserRole;
  is_active: boolean;
  last_login_at: string | null;
  created_at: string;
}